Eron portlamadi Everything is working correct
