--- app/src/main/java/uz/medion/ui/main/user/IOnBackPressed.kt
@@ -1,5 +1,0 @@
-package uz.medion.ui.main.user
-
-interface IOnBackPressed {
-    fun onBackPressed(): Boolean
-}
