+++ app/src/main/res/layout/activity_main.xml
@@ -15,10 +15,10 @@
         android:layout_width="match_parent"
         android:layout_height="match_parent">
 
-        <!--        <include-->
-        <!--            layout="@layout/app_bar_main"-->
-        <!--            android:layout_width="match_parent"-->
-        <!--            android:layout_height="match_parent" />-->
+<!--                <include-->
+<!--                    layout="@layout/app_bar_main"-->
+<!--                    android:layout_width="match_parent"-->
+<!--                    android:layout_height="match_parent" />-->
 
         <androidx.constraintlayout.widget.ConstraintLayout
             android:layout_width="match_parent"
@@ -58,6 +58,7 @@
                         android:layout_marginStart="60dp"
                         android:textColor="@color/white"
                         android:textSize="16sp"
+                        android:textIsSelectable="false"
                         app:layout_constraintBottom_toBottomOf="parent"
                         app:layout_constraintTop_toTopOf="parent" />
