+++ app/src/main/res/layout/fragment_search_view.xml
@@ -6,7 +6,7 @@
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".ui.main.user.searchView.SearchViewFragment">
+        tools:context=".ui.main.user.search.SearchViewFragment">
 
 
             <androidx.core.widget.NestedScrollView
